Improve URLPopover code in accordance with a review received
